2015-02-07 Fixed multi-monitor setup. Fixed full screen switching. Modified display of objectives in Settlement Dashboard tab.
----------------------------------------------------------------------------
1. Multi-Monitor Setup

Issue: if MSP (in eclipse) starts out on the right monitor, it does not stay on this "active" monitor. Rather, it opens up on the left monitor and then spans across both monitors in a two-monitor computer system in eclipse. Also, the menu bar is hidden. On the other hand, if it starts out out on the left monitor, it stays on the left monitor. This is not being consistent. Note that in both cases, the right monitor is the "primary" monitor as defined in the windows os.

Cause: this seems to be an on-going issue in javaFX according to the discussions in Stackoverflow. JavaFX lacks a consistent way to detect which monitors in a multi-monitor workstation. If the MSP/Eclipse has been moved to another screen, MSP will still start out on the previous screen.

See   http://stackoverflow.com/questions/25714573/open-javafx-application-on-active-screen-or-monitor-in-multi-screen-setup/25714762#25714762 

Solution:
a. Adds a new class StartUpLocation in /swing/tool/ folder. This class detects which monitor screen that the mouse cursor is on and defines the detected screen as the "active" screen. Note: it assumes the mouse pointer hovers on the same screen and will not be from monitor to monitor.

b. Adds a new method setMonitor(Stage stage) in MainMenu.java. Have each stage (and circleStage) call this method in numerous places in MainMenu, CrewEditorFX and ScenarioConfigEditorFX.

c. for now, set width to 1366 and height to 768 when starting MainScene and set the borderPane and scene to this resolution when starting. 

Note: in future, use Main Menu's config option to select the desired monitor and resolution.

----------------------------------------------------------------------------
2. Toggling on/off Full Screen Mode

Issue: users is unable to use the keyboard shortcut Ctrl+F to toggle off from the full screen mode. That said, however, users can still use mouse to select "Settings" on the menu bar and select "Full Screen Mode" to toggle off. 

Cause: there is a conflict of between the JavaFX's setFullScreenExitKeyCombination() and MainSceneMenu's setOnAction(e -> { }) methods for full screen switching using Ctrl+F.

Solution: adds a global boolean isFullScreenCache in MainSceneMenu.java to record previous state to resolve this conflict.

----------------------------------------------------------------------------
3. Settlement's Objective

a. Changed display of settlement objective in TabPanelDashboard.java from combobox style to radio buttons style in a toggle group.