2016-06-26 Created provision for more than one user defined crew. Replaced with bigger pause/quotation icons. Fixed quotation popup width.
Updated webstart-pack200-impl from 1.0-beta-6 to 1.0-beta-8-SNAPSHOT in mars-sim-main's pom.xml.

NOTE: if users has two LCD screen, there still exists a confusion of which screen to load the main menu, main scene and quotation popup if the mouse happens to be moving off from the primary screen to the secondary screen. Will need to hack how to track this in future build.

------------------------------------------------------------------------------------------------
I. NEW
------------------------------------------------------------------------------------------------

1. Created provision for more than one user defined mission crew. e.g. alpha, beta, charlie, etc.  (still-in-work)
a. Added new class Crew and Member and commented out a number of old List<String> defined for alpha crew in PersonConfig.java.
b. Replaced any reference to alpha crew as the first crew on the list.
c. Modified numerous methods in PersonConfig and calling of these methods in UnitManager, CrewEditor and CrewEditorFX. 
d. Added the new "crew" attribute to each <person> defined in <person-list> in people.xml and people.dtd

Note: this code changes make way for future work on the scenario editor, in which user will have full control of who will be sent to Mars on each mission.

------------------------------------------------------------------------------------------------------------------
II. CHANGES
------------------------------------------------------------------------------------------------------------------

1. By default, the width of the notification popup varies, depending on the resolution of the primary screen (i.e. Screen.getPrimary().getBounds().getWidth()). The new fix in setWidth() in QNotification.java should keep its width relatively constant regardless the primary screen size.

2. Switched to 64x64 pixel pause/quotation icons. Centered the word PAUSE and the pause icon.

3. Added new folder for holding PNotification.java and MessagePopup.java