2016-09-20b Fixed IOException when saving sim. Added temp file for saving. Shrunk jar size.
1. Added a temp file in messages.properties and used it in saveSimulation() in Simulation.java for assembling uncompressed objects into a file prior to compression.

2. Optimized OrbitInfo.java, Weather.java and Mar.java.

3. Fixed the IOException when saving sim by commenting out the use of EmotionJSONConfig.java.

4. Deleted springboot related files in /resource/static/ and /resource/webfxml/ in mars-sim-ui in order to shrink jar size.

5. Significantly reduced the size of map1.jpg and globe1.gif in /resource/docs/help/ in mars-sim-ui and renamed name to map.jpg and globe.gif. 

Note: 
a. Switched to using static String for matching the names of these two files in getURLType() in BroswerJFX.java.
b. 'map' and 'globe' are two usable commands in the Help browser. c. One may type 'map' and 'globe' in the address bar and hit enter to load the image.