2016-09-28 Fixed the ability to mute and unmute both the background sound track and the sound effect at the same time.
----------------------------------------------------------------------------------------
I. NEW
----------------------------------------------------------------------------------------

1. Added a new instance of OGGSoundClip called backgroundSoundTrack in AudioPlayer.

----------------------------------------------------------------------------------------
II. IMPROVEMENTS/CHANGES
----------------------------------------------------------------------------------------

1. Made AudioPlayer run clearly on either JavaFX UI thread or Swing AWT thread.

----------------------------------------------------------------------------------------
III. FIXES
----------------------------------------------------------------------------------------

1. Fixed the inability to mute and unmute the sound effect with Ctrl-M or toggling Mute on the menu.
a. Added playSound() in MainDesktopPane.
b. Checked on the state of the background sound track to determine if the sound effect should be played.
c. Modified various methods in AudioPlayer to mute and unmute both the background track/sound effect.


Notes: 
a. For one has to mute and unmute BOTH the background sound track and the sound effect at the same time.
b. In future, one should consider using separate control for (un)muting the background sound track and the sound effect.