2016-10-06a Enabled the option of changing the height of the MarsNet chat box. Added the keyword 'settlement','vehicle'/'rover' for the chat box conversion. Added Vehicle Statistics in chat box. Detected no speaker/headphone being plugged in. Cleaned up MainScene. Minor cosmetic changes.
----------------------------------------------------------------------------------------
I. NEW
----------------------------------------------------------------------------------------

1. Added options for changing the height of chatbox.

a. Used command /h1, /h2, /h3, /h4 in the chat box to set the height 
/h1 -- 256 pixels (by default)
/h2 -- 512 pixels
/h3 -- 768 pixels
/h4 -- 1024 pixels 

b. Set up the int field chatBoxHeight in MainScene for keeping track of the height of the chatBoxPane (which is a StackPane ) for the chat box.

2. Added the keyword "settlement" in the chat box if user is interested in inquiring about the settlement 

a. For now, it will only list how many settlements in total and the names of each settlement. 

3. Added querying # and types of vehicles/rovers (Parked or on mission) in a settlement by typing in 'vehicle' or 'rover' in askQuestion() in ChatBox.
a. Created a new VehicleType class--not fully put into use yet.
b. Created getExplorerRovers(), getLUVs(), getCargoRovers(), getTransportRovers(), getMissionVehicles() in Settlement and may be used for finding how many on mission and/or parked at settlement.


----------------------------------------------------------------------------------------
II. IMPROVEMENTS/CHANGES
----------------------------------------------------------------------------------------

1. Cleaned up MainScene. Minor cosmetic changes.

2. Sped up the slide down action of the menu bar in BorderSliderBar to avoid it from displaying empty content.

3. If users have no speaker or headphone plugged into the computer, OGGSoundClip will throw a number of exceptions--good that they are being caught. It should discern if speaker/headphone is being plugged instead of throwing all kinds of exceptions. Will return to see if there's better way for handling it. 

a. As an experiment, I converted nearly all throw new IOException() type of statements in OGGSoundClip and AudioPlayer to logger.log(SEVERE, ...) statement. 

b. For now, the log statement will simply say "no speaker/headphone detected" instead of throwing exceptions in the stack trace.


----------------------------------------------------------------------------------------
III. FIXES
----------------------------------------------------------------------------------------

None.