2016-10-07 Fix loading saved sim in headless. Prevent hanging if loading is unsuccessful. Avoid StackOverflowError. Clarify different mission cases in endMission() in VehicleMission.
----------------------------------------------------------------------------------------
I. NEW
----------------------------------------------------------------------------------------

None

----------------------------------------------------------------------------------------
II. IMPROVEMENTS/CHANGES
----------------------------------------------------------------------------------------

1. Remove splash.png in mars-sim-ui's \resources\images folder to save ~ 1MB. 

2. Define clearly the order of switches/arguments when typing in the input string in command prompt/terminal to run mars-sim,
'headless' must be ahead of 'load' if a user wants to load a particular saved sim.

e.g. java -jar mars-sim-main-3.1.0.jar headless load abcde.sim
 
In addition, 'abcde.sim' must be in the same directory as mars-sim-main-3.1.0.jar.
 
If default.sim is used, then the order of the switches doesn't matter.

3. Revise the order of arguments/switches in the eclipse launchers for MarsProjectFX.

4. Clean up the clutter in MainMenu.

5. Add notes to clarify different mission cases in a few if-else branching within endMission() in VehicleMission.

6. Tune the effect of stress/fatigue/hunger modifers on the performance reduction in recalculate() in PhysicalCondition.

a. Smoothen the abrupt impact of them on the performance.

b. Reduce the initial values of stress/fatigue/hunger at the start of the sim.

7. Revise README_V1.3.0.txt and CHANGELOG_V1.3.0.txt in mars-sim-main's /resource folder.

----------------------------------------------------------------------------------------
III. FIXES
----------------------------------------------------------------------------------------

1. Fix sequence of steps for calling handleLoadSimulation(argList) vs. handleLoadDefaultSimulation() in prepare() in MarsProjectFX.

2. Prevent hanging the console if user does NOT select a saved sim to load after opening FileChooser. 

-- Move startSimulation() from prepare() to inside the try catch of both handleLoadSimulation(argList) and handleLoadDefaultSimulation() to stop starting the sim if loading is proven unsuccessful.

3. Add capturing the user defined saved sim filename in main() in MarsProjectStarter.

-- The argument right behind the "load" argument in a type string is treated as the file name of the user defined saved sim. 

Note: to load this saved sim successfully in the command prompt, the sim file must be located in the same directory as the mars-sim's jar. 

4. Comment out setPhaseEnded(true) twice in endMission() in VehicleMission. Will perform tests to determine if setPhaseEnded(true) will cause NullPointerException in a loaded sim.

5. Avoid StackOverflowError from repetitive log statement on a person who are stranded in a building's airlock (due to high fatigue) by allowing walkToBed(accommodations, person, false) to fail.

a. Change from walkToBed(accommodations, person, false) to walkToBed(accommodations, person, true) in Sleep.

b. Annotate that calling getNewAction() enclosed inside a try-catch is still considered "experimental" in canExitAirlock() in ExistAirlock.

c. Add setPerformanceFactor() in within the try-catch phrase in part b above to "boost" a person's morale. Add 5% to the performance.  

Note 1: Assume a person will call up someone in intercomm to request help or to get out of any imminent danger/trouble Upon speaking to someone, his morale and performance level should shoot up. 

Note 2: Suggest to create a brand new Desire class (similar in scope to Preference class) for modifying which Task to take in short term.