2016-10-14b Allow tweaking the error margins for rovers going on a mission journey via in a xml.
Note 1 : this is in response to one user's last year comment that it was very unreasonable for a 4-settler settlement to send 3 folks out on a mineral/regolith mission for a location 1000+ miles away.

Note 2 : consider this just the first step toward allowing the user to directly change the mission range error margin in the simulation in real time in near future.


I. NEW
----------------------------------------------------------------------------------------

1. Allow changing the error margins for rovers on a mission in settlements.xml. 

- Rename FUEL_RANGE_ERROR_MARGIN in Vehicle to "fuel_range_error_margin" 

- Rename LIFE_SUPPORT_RANGE_ERROR_MARGIN in Rover to "life_support_range_error_margin"

- Add <mission-control> element in settlements.xml.

- Add <rover-life-support-range-error-margin value="3.0" /> element in settlements.xml.

- Add loadMissionControl() in SettlementConfig.

- Add getErrorMargin() in Rover and Vehicle.

II. IMPROVEMENTS/CHANGES
----------------------------------------------------------------------------------------

None.


III. FIXES
----------------------------------------------------------------------------------------

None.