2016-11-22 Switch to using enum for saving the location state of an unit. Improve/fix tracking down the location of a person/robot/vehicle/equipment in LocationTabPanel.
# NEW 

Add the new LocationStateType class.

Add 6 LocationStateType in messages.properties.

Add InsideVehicleOutsideOnMars (not used now).

Add getType() in all 7 location state classes.(not used now).


# FIXES

Revise updateLocation() in LocationTabPanel to properly track down the whereabout of a vehicle and an equipment.


# CHANGES

Abandon the use of LocationState and adopt the use of LocationStateType

Rename InsideVehicle to InsideVehicleInSettlement

Revise updatePersonState()

Rename getSettlement() to getParkedSettlement() for clarity (affects a lot of java classes)

Change getParkedSettlement() to read only containerUnit, not the top container unit