2017-02-15a Fix the cases when a user uses console to start mars-sim and inadvertently enter the switch with the dash ("-") in front of it. Fix the issue of dark background in MainDesktopPane. Add github issues page in about.html.
Note: normally, the symbol dash ("-") is NOT needed before the switch. If given, it will cause the switch to be bypassed and not captured.

e.g. "java -jar mars-sim.jar help" is valid.

e.g. "java -jar mars-sim.jar -help" used to be invalid but this commit to render it valid as well.