Fix suffocation. Add ResourceUtil
r3930
2017-04-10

Note 1: this commits further resolves the suffocation issue at https://github.com/mars-sim/mars-sim/issues/34

# CHANGES
1. Remove static modifier from the checkEVAProblem() in 
   EVAOperation.
2. Convert a few throw new IllegalStateException() statements 
   to logger.severe();
3. Remove static modifier when declaring AmountResource instances in
   Settlement, EVASuit and EVAOperation or it results in generating
   a lot of null and people will suffocate.

# NEW
1. Add ResourceUtil class for saving amount resources set and maps 
   that used to reside in AmountResource and AmountResourceConfig.

Note 3 : ResourceUtil is a singleton and it replaces the burden of
   using AmountResource for housing findAmountResource()--the
   method frequently called in mars-sim for obtaining the object
   reference of an amount resource.
   
2. Add newInstance() method in AmountResource.
3. Add checkEVAProblem() in WalkOutside so that it doesn't have to 
   call the checkEVAProblem() in EVAOperation.

# Fix
1. Modify clearTask() to avoid NullPointerException.
2. Add checking for empty farmBUildings to prevent 
   NullPointerException in getAvailableGreenhouse() in 
   TendGreenhouse. 

Note 2: after reloading the saved sim, an instance of  
amountResoureceConfig in ResourceUtil from the saved sim will be 
copied over here. 

This way, the newly created amountResourceConfig will be overridden 
and the resources from the original version of amountResourceConfig
will be preserved. 

The drawback is that this will create issues if one loads from a 
saved sim that has an outdated list of Amount Resources.