Fix crop issues/errors
r4025
2017-05-29

NOTE: impose the new mandate that when a greenhouse pick a new crop, 
it will automatically choose one that is NOT currently being planted --
i.e. "non-repeating" -- with the exception when the user specifically
chooses a crop of interest in the crop queue. This ensures a wider
variety of crop food is available. May allow user to tweak
how this restriction is imposed via a few checkboxes in near future.

# FIX
1. Fix numerous issues/errors caught by codacy.
   - Revise addWork(), requiresWork() in Farming and Crop.
2. Revise getRandomCropType() to obtain a random crop in Farming.
3. Limit the value of memory[0] in Crop to fix low health due to
   low sunlight when in darkness.
4. Revise getNewCrop() in Farming.
5. Fix the use of List<String> plantedCrops in Farming for keeping
   track of the list of crop that are currently being planted in 
   each greenhouse. 

# NEW 
1. Add containCrop() in Farming.

# CHANGE
1. Switch to me.tatarka's retrolambda (used by jfoenix)