Refine battery charging/discharge
r4090
2017-08-15

Note: the current charging/discharging model does not accurate account 
for the changing value of internal/load resistances at specific state of 
charge. 

In near future, we will portrait more accurately the internal
resistances 
of the grid batteries as well as the load resistances of each device or 
building to be powered during battery charging/discharging.


# CHANGE 
1. Refactor retrieveStoredEnergy(), updatePowerFlow(), and 
   storeExcessPower() in PowerGrid.
2. Refactor setEnergyStored(), in PowerStorage.
3. Change BATTERY_MAX_VOLTAGE to 375 V.
4. Rename NuclearHeatSource to ThermalNuclearSource.
5. Add Thermal Nuclear Source in HeatSourceType.
6. Convert calling AmountResource.findResource() to loading predefined
   AmountResource instances in ResourceUtil.

# ADD 
1. Define R_LOAD, degradationRatePerSol, systemEfficiency,
chargeBattery(), 
   dischargeBattery() in PowerGrid.
2. Define r_total, r_cell, numModules, cellsPerModule, C_rating, 
   getCRating(), updateHealth() and reconditionBattery() in
PowerStorage.