Add new maven submodule mars-sim-headless
r4297
2018-06-14

Note: this build is in response to shirishag75's comments in issue
#https://github.com/mars-sim/mars-sim/issues/66
and issue #https://github.com/mars-sim/mars-sim/issues/68 in that
openjfx may not be readily available in certain distro such as debian.


# NEW
1. Add mars-sim-headless to take advantage of Java 9 modularity in that
   mars-sim can be run in a console/terminal on a machine without
   the need of having openjfx/JavaFX package installed.

Note : mars-sim may currently run in a command line environment without
invoking the GUI. However, the mars-sim-main module has been structured
with a dependency on javafx/openjfx package and thus requires it to
start the simulation. On this ground, it's justifiable to separate the
codes into its own submodule that will enable mars-sim to fully compile
into a smaller binary package that run in purely headless mode.