Fix areocentric solar longitude and r
r4382
2018-06-30

Note: this commit is partially in response to the issue
https://github.com/mars-sim/mars-sim/issues/23 in that the areocentric
solar longitude (L_s) for Mars needs to be correlated with any given
Earth date so as to add future scenarios with any first Mars landing
date in mars-sim.

# FIX
1. Based on NASA Goddard (NISS)'s equations implemented in ClockUtils :
a. Compute the correct value of L_s at 2043-09-30 00:00:00 UTC to be
   12.7201, instead of being 252.5849 as used in the past.
b. Compute the correct value of r (the instantaneous distance between
   sun and Mars) at 2043-09-30 00:00:00 UTC to be 1.5876, instead of
   being 1.3818 or 1.6657 as used in the past.

# CHANGES
1. Adopt the use of the NISS equations for calculating L_s in
   - updateTimeLabels() in MainScene.
   - computeL_s() in MarsClock.