Improve skill tracking
r5017
2019-09-06

Note 1: experience points in Skill are non-cumulative.
It will reset back to zero at the next level.
It's simpler for calculation, tracking and observation of
how much experience points are needed.

Note 2: now even if a person has a skill with level 0, it
would show its level and needed experience points as well.

Note 3: the labor time is cumulative. It's currently in the
unit of millisols. The number can get very big over time.
May switch to a different unit of measurement later if
needed.

## NEW
1. Add labor time tracking for each skill.

## CHANGE
1. Add showing the experience points of a person/robot in
   PersonRobotChatUtils.
2. Modify SkillManager, Skill and TabPanelSkill

2. Modify MockSettlement.

## FIX
1. Fix unable to show new skills in TabPanelSkill.