Change the Function class to be Temporal.
This is a large change since the Function subclasses have a lot of
redundant and duplicate code. This has been refactored to provide
default implementation within the base Function class; hence reducing
the code in the subclasses.
Many of the Bugs indentified by SonarQube have been removed in the
reworking. Particularly the overloading of variables by the subclasses.
The Farming subclass has been significantly changed to utilise the
DataLogger pattern to hold the various rolling metrics.