Initial rework of MasterClock
mars-sim/mars-sim#287
This initial version of the change alters how time is managed. The
previous approach used real time as the key the mars updates constant
but the pulses could vary.
THis new approach takes the opposite where the pulse are nerly fixed and
each pulse varies the elapsed simulation time according to the ratio and
elapsed time.
There is still some tidying up to do.