Changes to track down people getting stuck walking
In long simulation runs people are getting stuck in AirLock Egress with
the inner door locked. I think the problem relates to Airlocks not being
closed probably after use. This means they eventually starve.
Sure this is not to do with the changes for applying time.
To help investigate there is now a DayDream task always available that
identifies People with nothing to do. The Mind.takeAction method is no
longer recursive as this is a potential memory overload problem. The net
effect is the full time should be used when a Task is changing in the
middle of a Pulse.