Resolved problem with infinite processing loop
close #290
The problem relates to some Tasks not processing time correctly:
* Tasks return a NaN as the remaining time
* Tasks never consume any time and hence become stuck
The `MindAction.takeAction` has been changed to compensate for these 2
cases and prevent them going into an infinite loop.
The AirLock Task is a problem sometimes.
The sleep is not directly the root cause but the MasterClock now allows
for a long elapsed time and calibrates.