Some Task.performTask are returning more time than available
Some tasks are increasing the remaining time for task execution by
returning a larger value than they are passed. This could throw the
impact the simulation performance as it "spins" rounding doing an ever
increasing available time.
Tending the greenhouse was one case where if a Crop is not fully tended
the time increases. This was the Crop.addWork method return the
remaining time not the consumed.