Thread safe around the global studies list.
Most logic has been moved to the Person class to reduce contention but
some shared use still exists. The global studies list is used by the
Task Meta layer to score potential Tasks which means it is heavily read.
But new Studies also add to the list.
Also the number of simulation threads can be increased to allow testing
with a higher number of threads. This fix has been tested with 4 threads
without problem.

Changed the CountryCommand to a wider ProfileCommand to help diagnose
what was happening.

#314