Avoid airlock depressurization conflict
r5694
2021-05-14

Note: this commit avoids an unique situation when the conflict
of EVA occupants arises. The airlock is ready for pressurization.
Suddenly a new member just entered Zone 1 and has yet donned on
the EVA suit, this new member will have to leave the airlock
right away in order to allow the airlock to proceed with
depressurization.

## CHANGE
1. Revise getNumOccupants() in Airlock.
2. Revise depressurizeChamber() in ExitAirlock.
3. Revise removeID() in Airlock.