Fix the Unit tests.
Problem is the FoodProductionConfig.destroy method is trying to
cleardown an immutable List. However the root cause is the UnitTest
repeatedly reload the configurations from XML. But as the configuration
classes are moving to be read-only they can be re-used in all tests with
problems.
Changed the loadConfig method to avoid destroying previously loaded
configurations.