OutOfMemory saving a simulation due to working space requirements
The OutOfMemory that occurs in the saving is not always due to Object
Serialization. The use of a ByteArray stream to hold the streamed data
consume a lot of memory that is not required.

The ObjectStrema can be streamed direct to/from the file via a GZip
stream. This significantly reduces the saving/loading time and removes
the need for a large block of memory to hold the intermediate version.

#335