Changed the order a researcher is removed from a Study
In theory the Study should be removed from the Person before the Person
is removed from the Study. A Person could be about to evaluate a
potential Study when they are removed by another thread because of
inactivity. This could lead to a race condition so flipping the order
will reduce the chance.

#357