MissionTableModel does not maintain a private copy of Missions
The MissionTableModel was directly sharing the mission list of the
MissionManager.
When a new mission was added it appears automatically in the model bit
the listener are not notified. One listener is the TableSorter which has
to update the row indexes which it fails to do. Therefore the two model
come misaligned with the SortModel not having all the row indexes which
generates a NPE creating blank cells.
#374