Implement reservation before egress airlock
r6195
2021-09-25

## ISSUE
1. https://github.com/mars-sim/mars-sim/issues/439
2. https://github.com/mars-sim/mars-sim/issues/340

NOTE : if a person has no reservation on an airlock, one
will not be allowed to enter the airlock for egress. For now,
this reservation is being tested on primarily DigLocalIce and
DigLocalRegolith task in conjunction with ExitAirlock. Will
allow a person to make reservation in other tasks in future.

## NEW
1. Add reservationMap, addReservation(), and removeReservation()
   in Airlock.
2. Add calling removeReservation() in enterAirlock() in
   ExitAirlock.
3. Add calling addReservation() in requestEgress() in
   ExitAirlock.
4. Add removing any expired entry in reservation map in
   timePassing() in Airlock.