Consider airlock chamber and/or reservation is full
r6224
2021-09-28

Note 0: this commit reduces the liklihood of arriving at an
airlock whose 4 chambers are full and/or reservation is full.

## ISSUE
1. #459
1. #439
2. #340


## CHANGE
1. Revise in getAirlock() in Settlement to consider :
   - Select airlock that fulfill either conditions:
   -- 1. chamber is not full
   -- 2. chamber is full but the reservation is not full

Note 1: getWalkableAvailableAirlock() in EVAOperation will call
getAirlock() in Settlement when finding an available airlock.

Note 2: Next is to save the result of calling this method so that
it will call once and use the same result for each frame,
instead of calling it a dozen times in each frame.