Rounding error in the Inventory Amount capacity calculations break UnitTest.
The problem is that there is a rounding issue with the Loading Unit Test
such that a Settlement can not be pre-loaded with the required Food
resource to support the test. The capacity exactly matches the needed
but it is loaded as 2 batches which means a rounding error occurs. This
results in the Inventory reporting no capacity.

Changes:
- The Loading Unit test adds an extra amount to the increase capacity to
cater for the rounding.
- Removed uncalled methods in Inventory class
- Correct scope of some public methods in the Inventory class
- The Stored & Capacity caches are not preloaded with all known
ResourcesIDs. The caches are laoded as resources are used. This reduces
the memory consumption and improves look-up speed.
- Inventory store & retrieve methods using AmountResource call the
equipment method using integer. Removes duplicate code.

#463