Fix problem with EVASuit reporting invalid pulse
The root cause is that the Settlement now only reference Equipment that
is within the Settlement; which is correct. But this causes a problem
with the time pulse since Equipment can be transferred out of the
Settlement and hence miss the pulse. Equipment are passive and should
not be pinged every pulse.

Secondly the EquipmentInventory could contain multiple references to the
same Equipment.

Changes:
- EquipmentInventory held equipment now modelled as a Set
- Fixed problem in EquipmentInventory returning an empty Bag that has
the wrong resource type; problem with DigLocal tasks
- Added a Vehicle refuel console expert command to fix Vehicles that run
out of fuel due to miscalculation in the distance.
- Fixed problem with ClassCast when a Person transfers out of a Building
into Vehicle
- Fixed problem with Storage not being loaded with initial resources.
This stops Missions starting as there was no methane.
-Fixed problem with Vehicle not being in Settlement when returning from
a Mission
- Fixed problem with Meteorology Field Study mission not adding specimen
boxes
- Settlement vehicle Console command redesigned to show vehicle status
instead of counts
- MissionSummary Console command takes an "active" argument to show just
active missions
- StructureResponse appendTableRow support Boolean arguments

#476