Fixed problem of Concurrent modification issue during Mission probability
The concurrent modification issue is due to the DeliveryUtil & TradeUtil
creating temporary Equipment in remote Settlements. This means that if
the remote Settlement is executing a pulse at that time there is a
chance of a concurrency problem.

Changes:
- DeliveryUtil & TradeUtil use the ContainerUtil to estimate capacity
- Create mission panel use the ContainerUtil instead of creating temp.
Equipment
- EquipmentInventory & MicroInventory remove the use of temp. ArrayLists
- Added method validateActiveSettlement to UnitManager allowing a Unit
to be compared against the active Settlement in the Thread.
- Fixed problem with Delivery missions not filling with enough fuel for
the return journey

#484
#483