Add a wait phase to EVA based Mission when there is poor sunlight
Changes:
- RoverMission.canStartEVA checks the if the sunlight is suitable to
start an EVA. If sunrise is within a reasonable time then the WAIT_PHASE
is entered; if not then TRAVELLING to the next NavPoint is started
- RoverMission has a new Phase called WAIT_SUNLIGHT. The descriptino
shows the expected sunrise
- performSWaitSunlight method check the current sunlight conditions to
see if an EVA can start. If the wait time is too long then the Mission
moves to the next TRAVELLING phase
- Added new method SurfaceFeatures.getSunrise that estimates the
MarsClock of sunrise at a location. This method is a rough estimate and
may need refining
- Fixed a problem with Collect Resource mission not collecting anything.
Problem with the capacity method being used.
- Changed the collection rate of ice collection.
- Changed number of container in CollectResource mission to make
reasonable collection targets.
- Fixed Person in a Vehicle always returns false for isInSettlement
- Fixed problem with CollectResourceMission having too large a
collection target. Now based on number of containers and sites.
- Re-enabled Consolidate containers in Vehicles but must be at least 3
partial containers present
- Changed the Vehicle 'refuel' command to generic 'refill' that takes
the resource to replenish as an argument, e.g. > refill oxygen

#469
#488