Fixed problem with crew getting stuck in parked Vehicle at end of mission
The problem is an UnloadEVA can result in someone being in the Vehicle
without means to refill their Suit because in parallel someone else
unloads the last of the Oxygen.

Changes:
- EVASuit has method to loadResources and unloadWaste. The former return
the percentage oxygen held
- EnterAirlock leaves any resources inside the EVA Suit; just unload
waste
- ExitAirlock uses the new EVASuit.loadResources to load EVASuit
- RoverMission in disembark loads 1 suit per person with resources
before UnLoad starts
- EVASuit has a new method to report how fully loaded it is
- RoverMission is now responsible for finding usable EVASuits; not
Vehicle
- RoverMission allows EVASuits that are not fully loaded to be re-used.
- Fixed problem in MeteorologyFieldStudy mission generating an
UnsupportOperation exception
- Stop Unlocking "stealing" oxygen from EVASuits; EquipmentInventory
only takes resources from Containers
- When a Vehicle is entered/leaves a Garage; the occupant are
transferred not just 'setContainerUnit'. Without this, Persons are still
registered as crew in the Vehicle when they are in the Garage.

#488