Convert WalkingSteps class to use Point2D instead of individual x & y.
This makes the migration to LocalPosition easier.
Also created a factory method in Walk. This removes the need to create
and validate the WalkingStep twice which currently happens. Current the
WalkingStpe is validated before creating the Walk Task which create the
WalkingSteps again. Expensive.
Now the WalkingSteps is only created once in the factory method.

Changes:

#470