Mission Tool correctly displayed collected resources
The root cause is a Barrel can only hold Liquid and not Solid as in Ice.
The Barrel in fact CAN hold LIQUID & SOLID so the Container/PhaseType
logic is changed so Barrels support both Liquid & Solid.

The approach is to rework how collected resources are reported and make
it generic instead of specialising.

Changes-
- ContainerUtil method isPhaseSupported replaces existing Phase method
to support multiples.
- GenericContainer allows ContainerUtil to decision the Phase
suitability.
- CollectResourceMission maintain a Map of the resources the mission
collects
- New method to return the collected resource amounts keyed on
resourceId
- CollectionResourceMissionPanel is generic driven by the resource Ids

close #517