Moved save logic from MasterClock to Simulation
The MasterClock should not know the logic of saving the Simulation; that
is the responsibility of the Simulation. The change involves performing
the save at the end of the pulse in the Simulation. The saving is
recorded in a pending save variable that is actioned at the end of the
pulse

Changes:
- MasterClock remove methods relating to saving Simulation
- Simulation provides a requestSave method that will trigger a save on
the next clock pulse.
- Simulation clockPulse method checks the save pending flag at the end
of the pulse and triggers a save if requested
- Simulation add a dedicated ClockListener to trigger auto saving; it
has a minimum duration set to the autosave period
- AutosaveScheduler is removed
- Last save datetime is held in the Simulation now instead of
EarthClock.
- MainWindow, ToolToolBar & SaveCommand use the new Simulation
requestSave method
- EnterAirlock checks if a Person is already present compensates for an
unknown problem in the Airlock logic