Improved the duration of EVAs in Missions
There are a few problems with the EVA phase being cut short.
From the Mission Log it is possible to see when an EVA is cut short due
to lack of light. And then how long it takes before the Vehicle resumes
travelling, i.e. time it take for Persons to return.

Part of #517 & #629

Changes:
- EVAOperation & EVAMission used the same meoth to assess EVA light
  suitability
- The EVA Light logic uses the original 1% of sunlight lower threshold
- RoverMission does not count persons in an EVA as in the Vehicle;
  these people may be in the process of leaving the Vehicle and hence
  become stranded
- Remove the Mission based rate calculations and solely rely on the
  individual Person rate.
- Add a Console Command called 'sunlight' for Vehicle, Person &
  Settlement that displays the current sunlight details at that
  location.