Add prospecting skill
r6966
2022-07-16

Note: Areology vs. Prospecting. Prospecting will focus more on
the use of prospecting instrument to survey and look for
minerals and resources in a locale. Areology skill is more
academic in nature and has to do with the ability to understand
the environment and make connections based what is being
observed.

1. https://github.com/mars-sim/mars-sim/issues/629

1. Add prospecting in SkillType.
2. Add researchQualityMap in Research.

1. Remove unneeded building and vehicle instance in
   MedicalCare.