Commerce based on a sell & Buy liat.
This represent a significant chaneg to how trading is performed. The
logic is now based on each Settlement creating a list of Goods to sell
and a list to buy.

When creating a Deal, the intersection of the sellers Sell list and the
buyers Buy list defines the list of Goods that could be traded. The load
is then built by taking items where the by price is higher than the sell
price.

Changes:
- Class Shipment represent a load and the associated revenue.
- GoodsManager now holds a buy & sell list. The lists are recalculated
  periodically.
 - The list calculations are basic at the moment.
 - CommerceUtil now removes the redundant old methods.
 - Trade console command now sows the buy & sell list.
Part of #601 & #506